Move main idea vault logic out of models
